+++ apps/io-session-manager-internal/src/services/__tests__/session-service.test.ts
@@ -1,17 +1,42 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import * as E from "fp-ts/lib/Either";
 import * as TE from "fp-ts/lib/TaskEither";
+import * as RTE from "fp-ts/lib/ReaderTaskEither";
 import { FiscalCode } from "@pagopa/ts-commons/lib/strings";
+import { QueueClient } from "@azure/storage-queue";
+import { TableClient } from "@azure/data-tables";
 import {
   RedisClientTaskMock,
   RedisRepositoryMock,
+  mockDelLollipopDataForUser,
+  mockDelUserAllSessions,
+  mockGetLollipopAssertionRefForUser,
   mockUserHasActiveSessionsOrLV,
 } from "../../__mocks__/repositories/redis.mock";
 import { SessionService } from "../session-service";
+import {
+  AuthLockRepositoryMock,
+  mockGetUserAuthenticationLocks,
+  mockIsUserAuthenticationLocked,
+  mockLockUserAuthentication,
+  mockUnlockUserAuthentication,
+} from "../../__mocks__/repositories/auth-lock.mock";
+import { LollipopRepositoryMock } from "../../__mocks__/repositories/lollipop.mock";
+import {
+  InstallationRepositoryMock,
+  mockDeleteInstallation,
+} from "../../__mocks__/repositories/installation.mock";
+import { anUnlockCode, anotherUnlockCode } from "../../__mocks__/user.mock";
+import {
+  forbiddenError,
+  toConflictError,
+  toGenericError,
+} from "../../utils/errors";
+import { aNotReleasedData } from "../../__mocks__/table-client.mock";
 
 const aFiscalCode = "SPNDNL80R13C555X" as FiscalCode;
 
-describe("Session Serviuce", () => {
+describe("Session Service#userHasActiveSessionsOrLV", () => {
   beforeEach(() => {
     vi.clearAllMocks();
   });
@@ -42,5 +67,280 @@ describe("Session Serviuce", () => {
 
     expect(mockUserHasActiveSessionsOrLV).toHaveBeenCalledTimes(1);
     expect(result).toEqual(E.left(customError));
+  });
+});
+
+describe("Session Service#lockUserAuthentication", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  const deps = {
+    FastRedisClientTask: RedisClientTaskMock,
+    SafeRedisClientTask: RedisClientTaskMock,
+    RedisRepository: RedisRepositoryMock,
+
+    AuthLockRepository: AuthLockRepositoryMock,
+    AuthenticationLockTableClient: {} as TableClient,
+    LollipopRepository: LollipopRepositoryMock,
+    RevokeAssertionRefQueueClient: {} as QueueClient,
+    InstallationRepository: InstallationRepositoryMock,
+    NotificationQueueClient: {} as QueueClient,
+  };
+
+  it("should succeed to lock an user authentication", async () => {
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(result).toEqual(E.right(null));
+  });
+
+  it("should fail when redis is not available", async () => {
+    const anErrorMessage = "ERROR";
+    const expectedError = toGenericError(
+      `Could not establish connection to redis: ${anErrorMessage}`,
+    );
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )({
+      ...deps,
+      FastRedisClientTask: TE.left(Error(anErrorMessage)),
+    })();
+
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should fail when the lock can't be retrieved", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(
+      "Something went wrong while checking the user authentication lock",
+    );
+    mockIsUserAuthenticationLocked.mockReturnValueOnce(RTE.left(anError));
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockIsUserAuthenticationLocked).toHaveBeenCalledTimes(1);
+
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should fail and return conflict when a lock is already present", async () => {
+    const expectedError = toConflictError(
+      "Another user authentication lock has already been applied",
+    );
+    mockIsUserAuthenticationLocked.mockReturnValueOnce(RTE.right(true));
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+    expect(mockIsUserAuthenticationLocked).toHaveBeenCalledTimes(1);
+
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should return generic error when assertionref retrieval fails", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(anError.message);
+    mockGetLollipopAssertionRefForUser.mockReturnValueOnce(TE.left(anError));
+
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockGetLollipopAssertionRefForUser).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should return generic error when delete of lollipopdata fails", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(anError.message);
+    mockDelLollipopDataForUser.mockReturnValueOnce(TE.left(anError));
+
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockDelLollipopDataForUser).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should return generic error when delete of all sessions fails", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(anError.message);
+    mockDelUserAllSessions.mockReturnValueOnce(TE.left(anError));
+
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockDelUserAllSessions).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should return generic error when the installation clear fails", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(
+      `Cannot delete Notification Installation: ${anError.message}`,
+    );
+    mockDeleteInstallation.mockReturnValueOnce(RTE.left(anError));
+
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockDeleteInstallation).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(E.left(expectedError));
+  });
+
+  it("should return generic error when the lock action fails", async () => {
+    const anError = Error("ERROR");
+    const expectedError = toGenericError(anError.message);
+    mockLockUserAuthentication.mockReturnValueOnce(RTE.left(anError));
+
+    const result = await SessionService.lockUserAuthentication(
+      aFiscalCode,
+      anUnlockCode,
+    )(deps)();
+
+    expect(mockLockUserAuthentication).toHaveBeenCalledTimes(1);
+    expect(result).toEqual(E.left(expectedError));
+  });
+});
+
+describe("Session Service#unlockUserAuthentication", () => {
+  beforeEach(() => {
+    vi.clearAllMocks();
+  });
+
+  const deps = {
+    AuthLockRepository: AuthLockRepositoryMock,
+    AuthenticationLockTableClient: {} as TableClient,
+  };
+  const aValidRequest = {
+    params: { fiscal_code: aFiscalCode },
+    body: { unlock_code: anUnlockCode },
+  };
+  const aValidRequestWithoutUnlockCode = {
+    params: { fiscal_code: aFiscalCode },
+    body: {},
+  };
+
+  it.each`
+    title                                      | request
+    ${"request contains unlock code"}          | ${aValidRequest}
+    ${"request does NOT contains unlock code"} | ${aValidRequestWithoutUnlockCode}
+  `(
+    "should succeed releasing CF-unlockcode when user authentication is locked and $title",
+    async ({ request }) => {
+      mockGetUserAuthenticationLocks.mockReturnValueOnce(
+        RTE.of([
+          aNotReleasedData,
+          { ...aNotReleasedData, rowKey: anotherUnlockCode },
+        ]),
+      );
+
+      const result = await SessionService.unlockUserAuthentication(
+        request.params.fiscal_code,
+        request.body.unlock_code,
+      )(deps)();
+
+      expect(result).toEqual(E.right(null));
+
+      expect(mockGetUserAuthenticationLocks).toHaveBeenCalledWith(aFiscalCode);
+      expect(mockUnlockUserAuthentication).toHaveBeenCalledWith(
+        aFiscalCode,
+        "unlock_code" in request.body
+          ? [request.body.unlock_code]
+          : [anUnlockCode, anotherUnlockCode],
+      );
+    },
+  );
+
+  it.each`
+    title                                      | request
+    ${"request contains unlock code"}          | ${aValidRequest}
+    ${"request does NOT contains unlock code"} | ${aValidRequestWithoutUnlockCode}
+  `(
+    "should succeed releasing CF-unlockcode when $title and query returns no records",
+    // This can occur in cases where there is either no user authentication lock or when an invalid unlock code has been provided.
+    async ({ request }) => {
+      mockGetUserAuthenticationLocks.mockReturnValueOnce(RTE.right([]));
+      const result = await SessionService.unlockUserAuthentication(
+        request.params.fiscal_code,
+        request.body.unlock_code,
+      )(deps)();
+
+      expect(result).toEqual(E.right(null));
+
+      expect(mockGetUserAuthenticationLocks).toHaveBeenCalledTimes(1);
+      expect(mockUnlockUserAuthentication).not.toHaveBeenCalled();
+    },
+  );
+
+  it("should return Forbidden releasing CF-unlockcode when unlock code does not match", async () => {
+    // This can occur in cases where there is either no user authentication lock or when an invalid unlock code has been provided.
+    mockGetUserAuthenticationLocks.mockReturnValueOnce(
+      RTE.right([{ ...aNotReleasedData, rowKey: anotherUnlockCode }]),
+    );
+
+    const result = await SessionService.unlockUserAuthentication(
+      aValidRequest.params.fiscal_code,
+      aValidRequest.body.unlock_code,
+    )(deps)();
+
+    expect(result).toEqual(E.left(forbiddenError));
+
+    expect(mockGetUserAuthenticationLocks).toHaveBeenCalled();
+    expect(mockUnlockUserAuthentication).not.toHaveBeenCalled();
+  });
+
+  it("should return GenericError when an error occurred retrieving user authentication lock data", async () => {
+    mockGetUserAuthenticationLocks.mockReturnValueOnce(
+      RTE.left(Error("an Error")),
+    );
+    const expectedError = toGenericError(
+      "Something went wrong while checking the user authentication lock",
+    );
+
+    const result = await SessionService.unlockUserAuthentication(
+      aValidRequest.params.fiscal_code,
+      aValidRequest.body.unlock_code,
+    )(deps)();
+
+    expect(result).toEqual(E.left(expectedError));
+
+    expect(mockUnlockUserAuthentication).not.toHaveBeenCalled();
+  });
+
+  it("should return GenericError when an error occurred releasing authentication lock", async () => {
+    mockGetUserAuthenticationLocks.mockReturnValueOnce(
+      RTE.of([aNotReleasedData]),
+    );
+
+    mockUnlockUserAuthentication.mockReturnValueOnce(
+      RTE.left(Error("an Error")),
+    );
+    const expectedError = toGenericError(
+      "Error releasing user authentication lock",
+    );
+
+    const result = await SessionService.unlockUserAuthentication(
+      aValidRequest.params.fiscal_code,
+      aValidRequest.body.unlock_code,
+    )(deps)();
+
+    expect(result).toEqual(E.left(expectedError));
+
+    expect(mockGetUserAuthenticationLocks).toHaveBeenCalledTimes(1);
+    expect(mockUnlockUserAuthentication).toHaveBeenCalledTimes(1);
   });
 });