+++ apps/io-session-manager-internal/src/utils/encoding.ts
@@ -1,0 +1,3 @@
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const base64EncodeObject = (_: any): string =>
+  Buffer.from(JSON.stringify(_)).toString("base64");