+++ apps/io-session-manager-internal/CHANGELOG.md
@@ -1,5 +1,29 @@
 # io-session-manager-internal
 
+## 0.6.0
+
+### Minor Changes
+
+- d902890: Added Release Auth Lock endpoint
+
+## 0.5.0
+
+### Minor Changes
+
+- e1ab55b: Expose `lockUserSession` endpoint
+
+## 0.4.0
+
+### Minor Changes
+
+- a8e0ba9: Expose `unlockUserSession` endpoint
+
+## 0.3.0
+
+### Minor Changes
+
+- 4e6084a: Ported AuthLock endpoint from io-backend
+
 ## 0.2.0
 
 ### Minor Changes